perf: migrate async SD ISR to multi-block CMD18/CMD25 transfers
Replace per-sector CMD17/CMD24 single-block transfers with full
multi-block CMD18/CMD25 transfers in the ISR state machine. Each
request now issues one command and one DMA setup covering the entire
cluster/request, with hardware auto-CMD12 terminating the transfer.
Interleaving moves from sector-level to request-level (2 fast per 1
slow) via a simplified state machine (SECTOR_DONE removed, replaced
with REQUEST_DONE). Fast queue expanded 16→32, slow queue reduced 8→4
to better match actual FatFS serialization patterns.

Performance improved noticeably but occasional freezes and e078 errors
remain on song load — requires further investigation.

fix: resolve int32_t/int template deduction errors in audio_engine.cpp
(std::max and std::clamp calls with mismatched literal types) introduced
in 7bfa471e "feat: re-enable cpuDireness with real-time OSTM0 timing".

+++ src/deluge/processing/engines/audio_engine.cpp
@@ -469,7 +469,7 @@ void cullVoices(int32_t direness, int32_t numAudio, int32_t numVoice) {
 
 		// Hard cull: direness at max or very close (13-14)
 		if (direnessOver >= 3) {
-			int32_t num_to_cull = std::max(direnessOver >> 1, 1);
+			int32_t num_to_cull = std::max(direnessOver >> 1, (int32_t)1);
 
 			// leave at least MIN_VOICES
 			num_to_cull = std::min(num_to_cull, numAudio + numVoice - MIN_VOICES);
@@ -553,7 +553,7 @@ void updateDirenessPostRender() {
 		// Slow release: ~12.5% step toward measured value
 		smoothedDireness8x += (measured8x - smoothedDireness8x) >> 3;
 	}
-	smoothedDireness8x = std::clamp(smoothedDireness8x, 0, kMaxDireness << 3);
+	smoothedDireness8x = std::clamp(smoothedDireness8x, (int32_t)0, (int32_t)(kMaxDireness << 3));
 
 	// Set starting direness for next buffer
 	cpuDireness = smoothedDireness8x >> 3;
